Move lots of things to their own files, stub matchmaking
