+++ .gitignore
@@ -1,5 +1,5 @@
 config.json
-
 .vscode
-
 env
+venv/
+__pycache__/
