+++ requirements.txt
@@ -1,2 +1,3 @@
 discord
 bs4
+requests
